Move velux base entity to separate module
